# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: Latest Binary Distributions from nodejs.org
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of the responses to the client (Accepts)
produces:
  - application/json
  - text/html
  - text/plain
paths:
  /download/{channel}/{os}/{dist}/{arch}:
    x-swagger-router-controller: download
    get:
      summary: Download the latest distribution of node.js
      description: Download the latest binary distribution of node.js given the channel (LTS or current), operating system (windows/macos/linux/source) distribution (installer/binary) and architecture (32-bit/64-bit/ARMv6/ARMv7/ARMv8)
      operationId: download
      parameters:
        - name: channel
          in: path
          description: The download channel (LTS or current)
          required: true
          type: string
          enum:
            - lts
            - current
        - name: os
          in: path
          description: The operating system (windows or macos or linux)
          required: true
          type: string
          enum:
            - windows
            - macos
            - linux
        - name: dist
          in: path
          description: Distribution format of download (binary archive or installer)
          required: true
          type: string
          enum:
            - installer
            - binary
        - name: arch
          in: path
          description: Distribution architecture (binary archive or installer)
          required: true
          type: string
          enum:
            - 32-bit
            - 64-bit
            - ARMv6
            - ARMv7
            - ARMv8
      responses:
        '404':
          description: Download not found
          schema:
            $ref: '#/definitions/NotFound'
  /download/{os}/{arch}:
    x-swagger-router-controller: download
    get:
      summary: Download the latest lts binary distribution of node.js
      description: Identical to /download/lts/{os}/binary/{arch} - Download the latest LTS binary distribution of node.js given the operating system (windows/macos/linux/source) and architecture (32-bit/64-bit/ARMv6/ARMv7/ARMv8)
      operationId: download_os_arch
      parameters:
        - name: os
          in: path
          description: The operating system (windows or macos or linux)
          required: true
          type: string
          enum:
            - windows
            - macos
            - linux
        - name: arch
          in: path
          description: Distribution architecture (binary archive or installer)
          required: true
          type: string
          enum:
            - 32-bit
            - 64-bit
            - ARMv6
            - ARMv7
            - ARMv8
      responses:
        '404':
          description: Download not found
          schema:
            $ref: '#/definitions/NotFound'

  /link/{os}/{arch}:
    x-swagger-router-controller: download
    get:
      summary: Download the latest lts binary distribution of node.js
      description: Identical to /download/lts/{os}/binary/{arch} - Download the latest LTS binary distribution of node.js given the operating system (windows/macos/linux/source) and architecture (32-bit/64-bit/ARMv6/ARMv7/ARMv8)
      operationId: link_os_arch
      parameters:
        - name: os
          in: path
          description: The operating system (windows or macos or linux)
          required: true
          type: string
          enum:
            - windows
            - macos
            - linux
        - name: arch
          in: path
          description: Distribution architecture (binary archive or installer)
          required: true
          type: string
          enum:
            - 32-bit
            - 64-bit
            - ARMv6
            - ARMv7
            - ARMv8
      responses:
        '404':
          description: Download not found
          schema:
            $ref: '#/definitions/NotFound'
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Link'
  /link/{channel}/{os}/{dist}/{arch}:
    x-swagger-router-controller: download
    get:
      summary: Download the latest lts binary distribution of node.js
      description: Identical to /download/lts/{os}/binary/{arch} - Download the latest LTS binary distribution of node.js given the operating system (windows/macos/linux/source) and architecture (32-bit/64-bit/ARMv6/ARMv7/ARMv8)
      operationId: link
      parameters:
        - name: channel
          in: path
          description: The download channel (LTS or current)
          required: true
          type: string
          enum:
            - lts
            - current
        - name: os
          in: path
          description: The operating system (windows or macos or linux)
          required: true
          type: string
          enum:
            - windows
            - macos
            - linux
        - name: dist
          in: path
          description: Distribution format of download (binary archive or installer)
          required: true
          type: string
          enum:
            - installer
            - binary
        - name: arch
          in: path
          description: Distribution architecture (binary archive or installer)
          required: true
          type: string
          enum:
            - 32-bit
            - 64-bit
            - ARMv6
            - ARMv7
            - ARMv8
      responses:
        '404':
          description: Download not found
          schema:
            $ref: '#/definitions/NotFound'
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Link'

  /hello:
    # binds a127 app logic to a route
    x-swagger-router-controller: hello_world
    get:
      description: Returns 'Hello' to the caller
      # used as the method name of the controller
      operationId: hello
      parameters:
        - name: name
          in: query
          description: The name of the person to whom to say hello
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  NotFound:
    required:
      - message
    properties:
      message:
        type: string
  Link:
    required:
      - url
      - version
    properties:
      version:
        type: string
      url:
        type: string
  HelloWorldResponse:
    required:
      - message
    properties:
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string